+++ intake/templatetags/my_tags.py
@@ -10,3 +10,9 @@ def verbose_name(obj):
 @register.filter
 def verbose_name_plural(obj):
     return obj._meta.verbose_name_plural
+
+@register.filter
+def shorten(obj, length=100):
+    if obj[:length].endswith((' ', ',',)):
+        return obj[:length - 1].strip() + '…'
+    return ' '.join(obj[:length].split()[:-1]) + '…'